requirements.txt
flask
beautifulsoup4
requests
newsapi-python
python-dotenv 